Fix off-by-one error in span batching logic
(also, add a couple of type specifications that were helpful when I was
debugging).

The problem here was that we waited until we had 21 spans ready to send, then
sent them out in batches of 20, which basically meant we always ended up with a
batch of 20 and another batch with a single item.